feat: complete rebase to ts for user block
