+++ package.json
@@ -45,6 +45,9 @@
     ]
   },
   "dependencies": {
+    "@types/express": "^4.17.12",
+    "@typescript-eslint/eslint-plugin": "^4.25.0",
+    "@typescript-eslint/parser": "^4.25.0",
     "cross-env": "7.0.3",
     "dotenv": "9.0.0",
     "express": "4.17.1",
@@ -53,6 +56,7 @@
     "yamljs": "0.3.0"
   },
   "devDependencies": {
+    "@types/uuid": "^8.3.0",
     "chai": "4.3.4",
     "debug": "4.3.1",
     "dirty-chai": "2.0.1",
@@ -63,11 +67,15 @@
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "3.4.0",
     "husky": "4.3.8",
+    "i": "^0.3.6",
     "jest": "26.6.3",
     "jsdoc": "^3.6.7",
     "lint-staged": "10.5.4",
     "nodemon": "2.0.7",
+    "npm": "^7.15.0",
     "prettier": "2.2.1",
-    "supertest": "6.1.3"
+    "supertest": "6.1.3",
+    "ts-node": "^10.0.0",
+    "typescript": "^4.3.2"
   }
 }
